+++ federation-js/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 > The changes noted within this `vNEXT` section have not been released yet.  New PRs and commits which introduce changes should include an entry in this `vNEXT` section as part of their development.  When a release is being prepared, a new header will be (manually) created below and the appropriate changes within that release will be moved into the new section.
 
-- _Nothing to see here. Stay tuned._
+- Fix a bug that would prevent types with similar field arguments from being composed as value types [PR #1298](https://github.com/apollographql/federation/pull/1298)
 
 ## v0.33.8